Moved model in their own package
