+++ models/src/main/scala/it/pagopa/interop/catalogmanagement/model/CatalogAttribute.scala
@@ -1,0 +1,7 @@
+package it.pagopa.interop.catalogmanagement.model
+
+sealed trait CatalogAttribute
+final case class SingleAttribute(id: CatalogAttributeValue)      extends CatalogAttribute
+final case class GroupAttribute(ids: Seq[CatalogAttributeValue]) extends CatalogAttribute
+
+object CatalogAttribute
