+++ models/src/main/scala/it/pagopa/interop/catalogmanagement/model/CatalogAttributeValue.scala
@@ -1,0 +1,5 @@
+package it.pagopa.interop.catalogmanagement.model
+
+final case class CatalogAttributeValue(id: String, explicitAttributeVerification: Boolean)
+
+object CatalogAttributeValue
