+++ models/src/main/scala/it/pagopa/interop/catalogmanagement/model/CatalogAttributes.scala
@@ -1,0 +1,9 @@
+package it.pagopa.interop.catalogmanagement.model
+
+final case class CatalogAttributes(
+  certified: Seq[CatalogAttribute],
+  declared: Seq[CatalogAttribute],
+  verified: Seq[CatalogAttribute]
+)
+
+object CatalogAttributes
