+++ models/src/main/scala/it/pagopa/interop/catalogmanagement/model/CatalogDocument.scala
@@ -1,0 +1,14 @@
+package it.pagopa.interop.catalogmanagement.model
+
+import java.time.OffsetDateTime
+import java.util.UUID
+
+final case class CatalogDocument(
+  id: UUID,
+  name: String,
+  contentType: String,
+  prettyName: String,
+  path: String,
+  checksum: String,
+  uploadDate: OffsetDateTime
+)
