+++ models/src/main/scala/it/pagopa/interop/catalogmanagement/model/CatalogItem.scala
@@ -1,0 +1,15 @@
+package it.pagopa.interop.catalogmanagement.model
+
+import java.util.UUID
+
+final case class CatalogItem(
+  id: UUID,
+  producerId: UUID,
+  name: String,
+  description: String,
+  technology: CatalogItemTechnology,
+  attributes: CatalogAttributes,
+  descriptors: Seq[CatalogDescriptor]
+)
+
+object CatalogItem
