+++ models/src/main/scala/it/pagopa/interop/catalogmanagement/model/CatalogItemTechnology.scala
@@ -1,0 +1,6 @@
+package it.pagopa.interop.catalogmanagement.model
+
+sealed trait CatalogItemTechnology
+case object Rest extends CatalogItemTechnology
+case object Soap extends CatalogItemTechnology
+object CatalogItemTechnology
