--- src/main/scala/it/pagopa/interop/catalogmanagement/model/Convertable.scala
@@ -1,5 +1,0 @@
-package it.pagopa.interop.catalogmanagement.model
-
-trait Convertable[A] {
-  def toApi: A
-}
